Moved Ioc Table in its own file to structure a bit and have a better view of the schema definitions
