+++ source/app/blueprints/rest/manage/manage_ioc_types_routes.py
@@ -24,7 +24,7 @@ from app import db
 from app.datamgmt.case.case_iocs_db import get_ioc_types_list
 from app.datamgmt.manage.manage_case_objs import search_ioc_type_by_name
 from app.iris_engine.utils.tracker import track_activity
-from app.models.models import Ioc
+from app.models.iocs import Ioc
 from app.models.models import IocType
 from app.models.authorization import Permissions
 from app.schema.marshables import IocTypeSchema
