+++ source/app/business/iocs.py
@@ -20,7 +20,7 @@ from marshmallow.exceptions import ValidationError
 
 from app import db
 from app.iris_engine.access_control.iris_user import iris_current_user
-from app.models.models import Ioc
+from app.models.iocs import Ioc
 from app.datamgmt.case.case_iocs_db import add_ioc
 from app.datamgmt.case.case_iocs_db import case_iocs_db_exists
 from app.datamgmt.case.case_iocs_db import check_ioc_type_id
