+++ source/app/datamgmt/case/case_assets_db.py
@@ -35,7 +35,7 @@ from app.models.models import CaseEventsAssets
 from app.models.cases import Cases
 from app.models.models import Comments
 from app.models.models import CompromiseStatus
-from app.models.models import Ioc
+from app.models.iocs import Ioc
 from app.models.models import IocAssetLink
 from app.models.models import IocType
 from app.models.authorization import User
