+++ source/app/datamgmt/manage/manage_cases_db.py
@@ -62,7 +62,7 @@ from app.models.authorization import OrganisationCaseAccess
 from app.models.authorization import User
 from app.models.authorization import UserCaseAccess
 from app.models.authorization import UserCaseEffectiveAccess
-from app.models.models import Ioc
+from app.models.iocs import Ioc
 from app.models.cases import CaseProtagonist
 from app.models.cases import CaseTags
 from app.models.cases import CaseState
