+++ source/app/models/models.py
@@ -15,6 +15,7 @@
 #  You should have received a copy of the GNU Lesser General Public License
 #  along with this program; if not, write to the Free Software Foundation,
 #  Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
+
 import datetime
 import enum
 import uuid
@@ -158,14 +159,6 @@ alert_assets_association = Table(
     Column('alert_id', ForeignKey('alerts.alert_id'), primary_key=True),
     Column('asset_id', ForeignKey('case_assets.asset_id'), primary_key=True)
 )
-
-alert_iocs_association = Table(
-    'alert_iocs_association',
-    db.Model.metadata,
-    Column('alert_id', ForeignKey('alerts.alert_id'), primary_key=True),
-    Column('ioc_id', ForeignKey('ioc.ioc_id'), primary_key=True)
-)
-
 
 class CaseAssets(db.Model):
     __tablename__ = 'case_assets'
@@ -389,42 +382,6 @@ class CaseTemplateReport(db.Model):
     report_type = relationship('ReportType')
     language = relationship('Languages')
     created_by_user = relationship('User')
-
-
-class Tlp(db.Model):
-    __tablename__ = 'tlp'
-
-    tlp_id = Column(Integer, primary_key=True)
-    tlp_name = Column(Text)
-    tlp_bscolor = Column(Text)
-
-
-class Ioc(db.Model):
-    __tablename__ = 'ioc'
-
-    ioc_id = Column(BigInteger, primary_key=True)
-    ioc_uuid = Column(UUID(as_uuid=True), server_default=text("gen_random_uuid()"), nullable=False)
-    ioc_value = Column(Text)
-    ioc_type_id = Column(ForeignKey('ioc_type.type_id'))
-    ioc_description = Column(Text)
-    ioc_tags = Column(String(512))
-    user_id = Column(ForeignKey('user.id'))
-    ioc_misp = Column(Text)
-    ioc_tlp_id = Column(ForeignKey('tlp.tlp_id'))
-    custom_attributes = Column(JSON)
-    ioc_enrichment = Column(JSONB)
-    modification_history = Column(JSON)
-
-    case_id = Column(ForeignKey('cases.case_id'), nullable=True)
-
-    user = relationship('User')
-    tlp = relationship('Tlp')
-    ioc_type = relationship('IocType')
-    case = relationship('Cases')
-    assets = relationship('IocAssetLink', back_populates='ioc', cascade="delete")
-    events = relationship('CaseEventsIoc', back_populates='ioc', cascade="delete")
-    comments = relationship('IocComments', back_populates='ioc', cascade="all, delete")
-    alerts = relationship('Alert', secondary=alert_iocs_association, back_populates='iocs')
 
 
 class CustomAttribute(db.Model):
